bugs fixes

+++ Semester 2/Algorithm & Data Structures/Homeworks/Homework 2/problem2.cpp
@@ -12,7 +12,7 @@ void print(deque<float> const &deq, string sep);
 int main()
 {
     deque<float> A;
-    int input;
+    float input;
     do
     {
         cin >> input;
